added conftest file, split object tests into separate files
