+++ tests/objects/test_urdf_object.py
@@ -1,0 +1,22 @@
+import numpy as np
+import pybullet as pb
+import pytest
+from deformable_gym.envs.pybullet_tools import BulletSimulation
+from deformable_gym.objects.bullet_object import ObjectFactory, UrdfObject
+from numpy.testing import assert_array_almost_equal
+
+
+TEST_POS = np.array([0, 0, 1])
+TEST_ORN = np.array([0, 0, 0])
+
+
+def test_urdf_object_creation(simulation):
+    obj = UrdfObject("plane.urdf", world_pos=TEST_POS, world_orn=TEST_ORN, fixed=True,
+                     client_id=simulation.get_physics_client_id())
+    pose = obj.get_pose()
+    assert_array_almost_equal(pose, np.array([0, 0, 1, 1, 0, 0, 0]))
+
+    vertices = obj.get_vertices()
+    assert len(vertices) == 0
+
+    assert obj.get_id() == 0
